+++ Nerd_STF/Mathematics/Equation.cs
@@ -1,3 +1,3 @@
 ﻿namespace Nerd_STF.Mathematics;
 
-public delegate double Equation(double x);
+public delegate float Equation(float x);
